refactor: Update install.sh script to use relative paths
The install.sh script was updated to use relative paths instead of absolute paths. This change improves the portability of the script and makes it easier to run in different environments. The script now uses the $scriptDirectory variable to reference the PowerShell scripts, ensuring that they can be executed regardless of the current working directory.